requirements.txt
matplotlib==3.10.1
nltk==3.9.1
numpy==2.2.4
pillow==11.1.0
opencv-python==4.11.0.86
scikit-image==0.25.2
scipy==1.15.2
streamlit==1.44.1
torch==2.6.0
torchvision==0.21.0
